requirements.txt
pytest==8.3.2
pytest-html==4.1.1
selenium==4.23.1
webdriver-manager==4.0.2
python-dateutil==2.9.0.post0